pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["str_assistant"]

[project]
name = "str-assistant"
version = "0.1.0"
description = "STR Assistant — AI agent for Short-Term Rental investment analysis"
readme = "README.md"
requires-python = "==3.12.2"
dependencies = [
    "deepagents==0.4.12",
    "fastapi==0.135.2",
    "uvicorn==0.42.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "langchain-openai==0.3.27",
    "langchain-mcp-adapters==0.2.2",
    "langgraph-checkpoint-postgres==3.0.5",
    "psycopg[binary,pool]==3.3.3",
    "structlog==25.5.0",
    "pyyaml==6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "pytest-cov==6.2.1",
    "ruff==0.12.2",
    "httpx==0.28.1",
]

[project.scripts]
str-assistant = "str_assistant.src.main:main"

[project.urls]
Homepage = "https://github.com/reidbryant/str-assistant"
Repository = "https://github.com/reidbryant/str-assistant"

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["E", "W", "F", "I"]
lint.ignore = ["E501"]
